extract course-related things into web/modules/course
