+++ modules/web/modules/course-searcher/course-results-list.js
@@ -1,7 +1,7 @@
 // @flow
 import React, { Component } from 'react'
 import map from 'lodash/map'
-import InlineCourse from '../../components/inline-course'
+import InlineCourse from '../course/inline-course'
 import debug from 'debug'
 const log = debug('web:react')
