+++ modules/web/modules/course/course-title.js
