+++ modules/web/modules/course/index.js
@@ -1,0 +1,2 @@
+export { default as InlineCourse } from './inline-course'
+export { default as ModalCourse } from './modal-course'
