+++ modules/web/modules/course/inline-course.js
@@ -6,12 +6,12 @@ import filter from 'lodash/filter'
 import isNull from 'lodash/isNull'
 import map from 'lodash/map'
 
-import { IDENT_COURSE } from '../../object-student/item-types'
+import { IDENT_COURSE } from '../../../object-student/item-types'
 
-import List from './list'
+import List from '../../components/list'
 import CourseTitle from './course-title'
 import { buildDeptNum } from '../../school-st-olaf-college/deptnums'
-import Icon from './icon'
+import Icon from '../../components/icon'
 import ModalCourse from './modal-course'
 
 import './inline-course.scss'
