+++ modules/web/modules/student-picker/index.js
@@ -1,0 +1,1 @@
+export { default } from './student-picker-container'
