+++ modules/web/modules/student-picker/student-list-item.js
@@ -3,12 +3,12 @@ import cx from 'classnames'
 import Link from 'react-router/lib/Link'
 import groupBy from 'lodash/groupBy'
 import map from 'lodash/map'
-import { interpose } from '../../../../lib/interpose'
-import { sortStudiesByType } from '../../../../object-student'
+import { interpose } from '../../../lib/interpose'
+import { sortStudiesByType } from '../../../object-student'
 
-import Button from '../../../components/button'
-import Icon from '../../../components/icon'
-import { iosTrashOutline, iosArrowForward } from '../../../icons/ionicons'
+import Button from '../../components/button'
+import Icon from '../../components/icon'
+import { iosTrashOutline, iosArrowForward } from '../../icons/ionicons'
 
 import './student-list-item.scss'
