+++ modules/web/modules/student-picker/student-list-item.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/variables.scss';
-@import '../../../styles/mixins.scss';
+@import '../../styles/variables.scss';
+@import '../../styles/mixins.scss';
 
 .student-list-item {
     @include link-undecorated();
