+++ modules/web/modules/student-picker/student-list.js
@@ -4,7 +4,7 @@ import map from 'lodash/map'
 import sortBy from 'lodash/sortBy'
 import fuzzysearch from 'fuzzysearch'
 
-import List from '../../../components/list'
+import List from '../../components/list'
 import StudentListItem from './student-list-item'
 
 import './student-list.scss'
