+++ modules/web/modules/student-picker/student-list.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/mixins.scss';
+@import '../../styles/mixins.scss';
 
 .student-list {
 	@include card();
