+++ modules/web/routes/index/index.js
@@ -2,7 +2,7 @@ export default {
 	getComponents(location, cb) {
 		require.ensure([], () => {
 			cb(null, {
-				content: require('./containers/student-picker').default,
+				content: require('../../modules/student-picker').default,
 			})
 		}, 'student-picker.components')
 	},
